create BlockManager interface
