+++ mcplanes-impl/src/main/java/co/tantleffbeef/mcplanes/McPlanes.java
@@ -82,7 +82,7 @@ public class McPlanes extends JavaPlugin {
 
         // initialize resource manager now that client jar has been downloaded
         resourceManager = new ResourceManager(this, persistentDataKeyManager, webserverFolder, clientJar);
-        blockManager = new BlockManager(persistentDataKeyManager, getServer(), resourceManager);
+        blockManager = new MCPBlockManager(persistentDataKeyManager, getServer(), resourceManager);
 
         // // Listeners!!!
